1、将UniqueQueue提取出来 2、修改一些小bug
